requirements.txt
streamlit
pandas
duckdb
langchain
openai
tiktoken
plotly
seaborn
matplotlib
python-dotenv
langchain-community
fpdf
